requirements-vertex.txt
# Vertex AI specific requirements
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0
accelerate>=0.20.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.3.0

# Google Cloud
google-cloud-storage>=2.10.0
google-cloud-aiplatform>=1.30.0
gcsfs>=2023.6.0

# Experiment tracking
tensorboard>=2.13.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
requests>=2.31.0
Pillow>=10.0.0
tqdm>=4.65.0